+++ opensquirrel/default_instructions.py
@@ -17,11 +17,13 @@ from opensquirrel.ir.default_gates import (
     SWAP,
     X90,
     Y90,
+    Z90,
     CRk,
     H,
     I,
     MinusX90,
     MinusY90,
+    MinusZ90,
     Rn,
     Rx,
     Ry,
@@ -30,6 +32,7 @@ from opensquirrel.ir.default_gates import (
     SDagger,
     T,
     TDagger,
+    U,
     X,
     Y,
     Z,
@@ -56,8 +59,10 @@ default_bsr_without_params_set = {
     "Y": Y,
     "Y90": Y90,
     "Z": Z,
+    "Z90": Z90,
     "mX90": MinusX90,
     "mY90": MinusY90,
+    "mZ90": MinusZ90,
 }
 default_bsr_full_params_set: Mapping[str, type[SingleQubitGate]]
 default_bsr_full_params_set = {
@@ -69,54 +74,46 @@ default_bsr_with_angle_param_set = {
     "Ry": Ry,
     "Rz": Rz,
 }
-default_bloch_sphere_rotation_set: Mapping[str, type[SingleQubitGate]]
-default_bloch_sphere_rotation_set = {
+
+default_bsr_unitary_param_set: dict[str, type[SingleQubitGate]] = {
+    "U": U,
+}
+default_bloch_sphere_rotation_set: dict[str, type[SingleQubitGate]] = {
     **default_bsr_full_params_set,
     **default_bsr_without_params_set,
     **default_bsr_with_angle_param_set,
+    **default_bsr_unitary_param_set,
 }
-default_controlled_gate_set: Mapping[str, type[ControlledGate]]
-default_controlled_gate_set = {
+default_controlled_gate_set: dict[str, type[ControlledGate]] = {
     "CNOT": CNOT,
     "CR": CR,
     "CRk": CRk,
     "CZ": CZ,
 }
-default_matrix_gate_set: Mapping[str, type[MatrixGate]]
-default_matrix_gate_set = {
+default_matrix_gate_set: dict[str, type[MatrixGate]] = {
     "SWAP": SWAP,
 }
 default_gate_alias_set = {
     "Hadamard": H,
     "Identity": I,
 }
-
-default_gate_set: Mapping[str, type[Gate]]
-default_gate_set = {
+default_gate_set: dict[str, type[Gate]] = {
     **default_bloch_sphere_rotation_set,
     **default_controlled_gate_set,
     **default_matrix_gate_set,
     **default_gate_alias_set,
 }
-
-default_unitary_set: Mapping[str, type[Unitary]]
-default_unitary_set = {**default_gate_set}
-
-default_non_unitary_set: Mapping[str, type[NonUnitary]]
-default_non_unitary_set = {
+default_unitary_set: dict[str, type[Unitary]] = {**default_gate_set}
+default_non_unitary_set: dict[str, type[NonUnitary]] = {
     "init": Init,
     "measure": Measure,
     "reset": Reset,
 }
-
-default_control_instruction_set: Mapping[str, type[ControlInstruction]]
-default_control_instruction_set = {
+default_control_instruction_set: dict[str, type[ControlInstruction]] = {
     "barrier": Barrier,
     "wait": Wait,
 }
-
-default_instruction_set: Mapping[str, type[Instruction]]
-default_instruction_set = {
+default_instruction_set: dict[str, type[Instruction]] = {
     **default_unitary_set,
     **default_non_unitary_set,
     **default_control_instruction_set,