+++ opensquirrel/ir/semantics/bsr.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import cmath
 import math
-from math import pi
+from math import cos, floor, log10, pi, sin
 from typing import TYPE_CHECKING, Any, SupportsFloat
 
 import numpy as np
@@ -12,15 +12,13 @@ from opensquirrel.common import ATOL, normalize_angle, repr_round
 from opensquirrel.ir.expression import Axis, AxisLike, Float
 from opensquirrel.ir.ir import IRNode
 from opensquirrel.ir.semantics.gate_semantic import GateSemantic
+from opensquirrel.utils.general_math import acos
 
 if TYPE_CHECKING:
     from opensquirrel.ir import IRVisitor
-    from opensquirrel.ir.expression import Expression
 
 
 def bsr_from_matrix(matrix: ArrayLike | list[list[int | DTypeLike]]) -> BlochSphereRotation:
-    from opensquirrel.utils import acos
-
     cmatrix = np.asarray(matrix, dtype=np.complex128)
     assert cmatrix.shape == (2, 2)
     a, b, c, d = map(complex, cmatrix.flatten())
@@ -91,6 +89,51 @@ class BlochSphereRotation(GateSemantic, IRNode):
 
         return False
 
+    def __mul__(self, other: BlochSphereRotation) -> BlochSphereRotation:
+        """Computes the single qubit gate resulting from the composition of two single
+        qubit gates, by composing the Bloch sphere rotations of the two gates.
+        The first rotation (A) is applied and then the second (B):
+
+        As separate gates:
+            A q
+            B q
+
+        As linear operations:
+            (B * A) q
+
+        If the final single qubit gate is anonymous, we try to match it to a default gate.
+
+        Uses Rodrigues' rotation formula (see https://en.wikipedia.org/wiki/Rodrigues%27_rotation_formula).
+        """
+        acos_argument = cos(self.angle / 2) * cos(other.angle / 2) - sin(self.angle / 2) * sin(
+            other.angle / 2
+        ) * np.dot(self.axis, other.axis)
+        combined_angle = 2 * acos(acos_argument)
+
+        if abs(sin(combined_angle / 2)) < ATOL:
+            return bsr_from_matrix([[1, 0], [0, 1]])
+
+        order_of_magnitude = abs(floor(log10(ATOL)))
+        combined_axis = np.round(
+            (
+                1
+                / sin(combined_angle / 2)
+                * (
+                    sin(self.angle / 2) * cos(other.angle / 2) * self.axis.value
+                    + cos(self.angle / 2) * sin(other.angle / 2) * other.axis.value
+                    + sin(self.angle / 2) * sin(other.angle / 2) * np.cross(other.axis, self.axis)
+                )
+            ),
+            order_of_magnitude,
+        )
+
+        combined_phase = np.round(self.phase + other.phase, order_of_magnitude)
+        return BlochSphereRotation(
+            axis=combined_axis,
+            angle=combined_angle,
+            phase=combined_phase,
+        )
+
 
 class BsrNoParams(BlochSphereRotation):
     def __init__(
@@ -114,7 +157,7 @@ class BsrFullParams(BlochSphereRotation):
         self.phi = Float(self.phase)
 
     @property
-    def arguments(self) -> tuple[Expression, ...]:
+    def arguments(self) -> tuple[Float, ...]:
         return self.nx, self.ny, self.nz, self.theta, self.phi
 
     def accept(self, visitor: IRVisitor) -> Any:
@@ -133,9 +176,38 @@ class BsrAngleParam(BlochSphereRotation):
         self.theta = Float(self.angle)
 
     @property
-    def arguments(self) -> tuple[Expression, ...]:
+    def arguments(self) -> tuple[Float, ...]:
         return (self.theta,)
 
     def accept(self, visitor: IRVisitor) -> Any:
         visit_bsr = super().accept(visitor)
         return visit_bsr if visit_bsr is not None else visitor.visit_bsr_angle_param(self)
+
+
+class BsrUnitaryParams(BlochSphereRotation):
+    def __init__(
+        self,
+        theta: SupportsFloat,
+        phi: SupportsFloat,
+        lmbda: SupportsFloat,
+    ) -> None:
+        bsr = self._get_bsr(theta, phi, lmbda)
+        BlochSphereRotation.__init__(self, bsr.axis, bsr.angle, bsr.phase)
+        self.theta = Float(theta)
+        self.phi = Float(phi)
+        self.lmbda = Float(lmbda)
+
+    @property
+    def arguments(self) -> tuple[Float, ...]:
+        return self.theta, self.phi, self.lmbda
+
+    @staticmethod
+    def _get_bsr(theta: SupportsFloat, phi: SupportsFloat, lmbda: SupportsFloat) -> BlochSphereRotation:
+        a = BlochSphereRotation((0, 0, 1), lmbda, 0)
+        b = BlochSphereRotation((0, 1, 0), theta, 0)
+        c = BlochSphereRotation((0, 0, 1), phi, (float(phi) + float(lmbda)) / 2)
+        return a * b * c
+
+    def accept(self, visitor: IRVisitor) -> Any:
+        visit_bsr = super().accept(visitor)
+        return visit_bsr if visit_bsr is not None else visitor.visit_bsr_unitary_params(self)