+++ opensquirrel/writer/writer.py
@@ -27,6 +27,7 @@ from opensquirrel.ir.semantics import (
     BsrAngleParam,
     BsrFullParams,
     BsrNoParams,
+    BsrUnitaryParams,
     ControlledGate,
     MatrixGate,
 )
@@ -108,6 +109,12 @@ class _WriterImpl(IRVisitor):
     def visit_bsr_angle_param(self, gate: BsrAngleParam) -> str:
         theta_argument = gate.theta.accept(self)
         return f"({theta_argument})"
+
+    def visit_bsr_unitary_params(self, gate: BsrUnitaryParams) -> str:
+        theta_argument = gate.theta.accept(self)
+        phi_argument = gate.phi.accept(self)
+        lmbda_argument = gate.lmbda.accept(self)
+        return f"({theta_argument}, {phi_argument}, {lmbda_argument})"
 
     def visit_swap(self, gate: SWAP) -> Any:
         qubit_operand_0 = gate.qubit_0.accept(self)